+++ src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/_constants.py
@@ -4,7 +4,7 @@
 # --------------------------------------------------------------------------------------------
 
 NODE_VERSION_DEFAULT = "10.14"
-NETCORE_VERSION_DEFAULT = "2.0"
+NETCORE_VERSION_DEFAULT = "2.1"
 DOTNET_VERSION_DEFAULT = "4.7"
 PYTHON_VERSION_DEFAULT = "3.7"
 NETCORE_RUNTIME_NAME = "dotnetcore"
@@ -15,7 +15,7 @@ OS_DEFAULT = "Windows"
 STATIC_RUNTIME_NAME = "static"  # not an official supported runtime but used for CLI logic
 NODE_VERSIONS = ['4.4', '4.5', '6.2', '6.6', '6.9', '6.11', '8.0', '8.1', '8.9', '8.11', '10.1', '10.10', '10.14']
 PYTHON_VERSIONS = ['3.7', '3.6', '2.7']
-NETCORE_VERSIONS = ['1.0', '1.1', '2.0', '2.1', '2.2']
+NETCORE_VERSIONS = ['1.0', '1.1', '2.1', '2.2']
 DOTNET_VERSIONS = ['3.5', '4.7']
 LINUX_SKU_DEFAULT = "P1V2"
 RUNTIME_TO_IMAGE = {